requirements.txt
fastapi
fastapi-cli
mangum
mysql-connector-python
numpy
pandas
python-dotenv
uvicorn



